moved prebuilt download to its own module
